All_Project_Files.ZIP
As students at ECE Engineering School (Paris), we had to make a little gameboy with specifics components and an unknown PCB card designed by our school. We decided to make a Snake game, and used the external library OLED85.h write by "cultsauce" on GitHub. Everything is detailed in the report (in french).

Here are the results : codes, report and pictures/bitmaps designed...